+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Make pop-up window active explicity for some window systems.
 - Fix selected language on search.
 - Fix size of auto language buttons in pop-up window.
+- Linux: create autorun file from desktop file.
 - Windows: Load Qt library translations.
 - Windows: Improve getting selection.